tweaked a bit

+++ src/main/java/ds/assign1/accounts/services/AccountService.java
@@ -10,6 +10,7 @@ import ds.assign1.accounts.dtos.builders.FullAccountBuilder;
 import ds.assign1.accounts.entities.Account;
 import ds.assign1.accounts.entities.Credentials;
 import ds.assign1.accounts.repos.AccountRepo;
+import ds.assign1.login.LoginDTO;
 import ds.assign1.login.infrastructure.ILoginService;
 import ds.assign1.mapping.infrastructure.IAccountService;
 import lombok.RequiredArgsConstructor;
@@ -31,7 +32,7 @@ public class AccountService implements ILoginService, IAccountService {
 
     private final AccountValidators VALIDATORS;
 
-    public FullAccountDTO createAccount(AccountDTO accountDTO, CredentialsDTO credentialsDTO){
+    public LoginDTO createAccount(AccountDTO accountDTO, CredentialsDTO credentialsDTO){
         VALIDATORS.usernameValidators(credentialsDTO.getUsername());
         VALIDATORS.passwordValidator(credentialsDTO.getPassword());
 
@@ -41,7 +42,7 @@ public class AccountService implements ILoginService, IAccountService {
 
         createdAccount = accountRepo.save(createdAccount);
 
-        return FullAccountBuilder.build(createdAccount);
+        return new LoginDTO(createdAccount.getId(), createdAccount.getRole().toString());
     }
 
     public List<ReturnAccountDTO> getAccounts(){
@@ -62,6 +63,16 @@ public class AccountService implements ILoginService, IAccountService {
         }
 
         return AccountBuilder.buildReturn(foundAccount.get());
+    }
+
+    public FullAccountDTO findAccountDetails(UUID idToSearch){
+        Optional<Account> foundAccount = accountRepo.findById(idToSearch);
+        if(!foundAccount.isPresent()){
+            LOGGER.error("There's no such account with id {}", idToSearch);
+            throw new ResourceNotFoundException(Account.class.getSimpleName() + " with id " + idToSearch);
+        }
+
+        return FullAccountBuilder.build(foundAccount.get());
     }
 
     public UUID updateAccount(UUID idToUpdate, AccountDTO dto){
@@ -110,7 +121,12 @@ public class AccountService implements ILoginService, IAccountService {
             throw new ResourceNotFoundException(Account.class.getSimpleName() + " with id " + idToDelete);
         });
 
-        accountRepo.delete(account);
+        if(account.getDeviceList() == null){
+            accountRepo.delete(account);
+        }
+        else{
+            throw new RuntimeException("The account has devices paired with it");
+        }
 
         return account.getId();
     }
@@ -129,7 +145,7 @@ public class AccountService implements ILoginService, IAccountService {
 
     @Override
     public String getRole(UUID id) {
-        return findAccountById(id).toString();
+        return findAccountById(id).getRole().toString();
     }
 
     @Override
